pyproject.toml
[project]
name = "ai-girlfriend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ffmpeg>=1.4",
    "numpy>=2.2.5",
    "openai>=1.79.0",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "speechrecognition>=3.14.3",
]
